# Description

This PR improves the runtime performance of the `reindex_kv_cache` and `copy_to_kv_cache` kernels, by remapping CUDA blocks/threads to different `b`, `s`, `hxd` dimensions, parallelizing one loop that was previously serial, and skipping reindexing if there are no new sequences added to the batch.

Example problem size: b=64, max_s=1024, current_s=128, new_s=8, h=12, d=256, on Hopper
Before: reindex: 4.8ms, copy: 1.3ms
After: reindex: 46us, copy: 13us

## Type of change

- [ ] Documentation change (change only to the documentation, either a fix or a new content)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Infra/Build change
- [ ] Code refactoring

## Changes

Please list the changes introduced in this PR:

- Optimize reindex and copy kernels in KV caching

# Checklist:

- [x] I have read and followed the [contributing guidelines](https://github.com/NVIDIA/TransformerEngine/blob/main/CONTRIBUTING.rst)
- [x] The functionality is complete
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes